Update immediate action demo screen to push instead of using modal
