+++ example/src/screens/EmbeddedPaymentElementImmediateActionScreen.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Alert, View, Text, Modal } from 'react-native';
+import { View, Text } from 'react-native';
 import Button from '../components/Button';
 import PaymentScreen from '../components/PaymentScreen';
 import CustomerSessionSwitch from '../components/CustomerSessionSwitch';
@@ -23,7 +23,6 @@ export default function EmbeddedPaymentElementImmediateActionScreen() {
   const navigation = useNavigation();
 
   // Local UI state
-  const [modalVisible, setModalVisible] = React.useState(false);
   const [customerKeyType, setCustomerKeyType] = React.useState<
     'legacy_ephemeral_key' | 'customer_session'
   >('legacy_ephemeral_key');
@@ -209,35 +208,19 @@ export default function EmbeddedPaymentElementImmediateActionScreen() {
     },
     [customerKeyType]
   );
-  const [loading, setLoading] = React.useState(false);
 
   // Hook into Stripe element
-  const {
-    embeddedPaymentElementView,
-    paymentOption,
-    confirm,
-    clearPaymentOption,
-    loadingError,
-  } = useEmbeddedPaymentElement(intentConfig!, elementConfig!);
-
-  // Payment action
-  const handlePay = React.useCallback(async () => {
-    setLoading(true);
-    const result = await confirm();
-    if (result.status === 'completed')
-      Alert.alert('Success', 'Payment confirmed');
-    else if (result.status === 'failed')
-      Alert.alert('Error', `Failed: ${result.error.message}`);
-    else Alert.alert('Cancelled');
-    setLoading(false);
-  }, [confirm]);
+  const { embeddedPaymentElementView, confirm, loadingError } =
+    useEmbeddedPaymentElement(intentConfig!, elementConfig!);
 
   return (
     <PaymentScreen
       onInit={() => {
         initialize(() => {
           console.log('Immediate Action callback called.');
-          setModalVisible(false);
+          navigation.navigate('EmbeddedPaymentElementConfirmScreen', {
+            confirm: confirm,
+          });
         });
       }}
     >
@@ -247,68 +230,24 @@ export default function EmbeddedPaymentElementImmediateActionScreen() {
           setCustomerKeyType(val ? 'customer_session' : 'legacy_ephemeral_key')
         }
       />
-      <View style={{ flexDirection: 'row', gap: 20, marginBottom: 10 }}>
-        <Button
-          variant="default"
-          title="Open screen"
-          onPress={() => {
-            navigation.navigate('HomeScreen');
-          }}
-        />
-        <Button
-          variant="default"
-          title="Open modal"
-          onPress={() => {
-            setModalVisible(true);
-          }}
-        />
-      </View>
-      <View style={{ paddingVertical: 16 }}>
-        <Text style={{ fontSize: 16, fontWeight: '600' }}>
-          {paymentOption?.label ?? 'No option'}
-        </Text>
-      </View>
-
-      <Button
-        variant="primary"
-        title="Pay"
-        onPress={handlePay}
-        loading={loading}
-        disabled={!paymentOption}
-      />
-
       <Button
         variant="default"
-        title="Clear"
-        onPress={clearPaymentOption}
-        disabled={!paymentOption}
-      />
-      <Modal
-        visible={modalVisible}
-        animationType="slide"
-        onRequestClose={() => {
-          setModalVisible(false);
+        title="Open screen"
+        onPress={() => {
+          navigation.navigate('HomeScreen');
         }}
-      >
-        <View style={{ padding: 20 }}>
-          {loadingError && (
-            <View style={{ padding: 12, backgroundColor: '#fee', margin: 8 }}>
-              <Text style={{ color: '#900', fontWeight: '600' }}>
-                Failed to load payment form:
-              </Text>
-              <Text style={{ color: '#900' }}>{loadingError.message}</Text>
-            </View>
-          )}
-          {embeddedPaymentElementView}
-          <Button
-            variant="default"
-            title="Close modal"
-            onPress={() => {
-              setModalVisible(false);
-            }}
-          />
-        </View>
-      </Modal>
+      />
+      <View style={{ padding: 20 }}>
+        {loadingError && (
+          <View style={{ padding: 12, backgroundColor: '#fee', margin: 8 }}>
+            <Text style={{ color: '#900', fontWeight: '600' }}>
+              Failed to load payment form:
+            </Text>
+            <Text style={{ color: '#900' }}>{loadingError.message}</Text>
+          </View>
+        )}
+        {embeddedPaymentElementView}
+      </View>
     </PaymentScreen>
   );
 }
